Created router stretch endpoint at /expense, returns budget diff
Created the endpoint at /expense that returns the difference between the budget and the expense.

To move it to the correct endpoint wouldnt be a problem though as you would just do the same thing. 

I did have an issue with reassigning Expense after it is compiled when I imported it into the Budget router and I wasnt able to solve that.